requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Cache & Redis
redis[hiredis]==5.0.1
aioredis==2.0.1

# Task Queue
celery[redis]==5.3.4

# Environment & Configuration
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# File handling
aiofiles==23.2.1

# Date/Time utilities
python-dateutil==2.8.2
pytz==2023.3

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9

# Search (OpenSearch/Elasticsearch client)
opensearch-py==2.4.2

# WebSocket support (already included in fastapi/uvicorn)

# Optional: Background tasks
dramatiq[redis]==1.15.0